+++ companion/hid_listener/Console.py
@@ -1,0 +1,17 @@
+
+from abc import ABC
+from .Device import Device
+
+
+class Console(ABC):
+    @staticmethod
+    def print_connect(device: Device) -> None:
+        pass
+
+    @staticmethod
+    def print_disconnect(device: Device) -> None:
+        pass
+
+    @staticmethod
+    def print_couldnt_connect(device: Device) -> None:
+        pass
