+++ companion/hid_listener/DeviceListener.py
@@ -1,85 +1,16 @@
 
 from threading import Thread
-from typing import TypedDict
 from time import sleep
-from abc import ABC
 
-from message_observers import Observer
+from .observers import Observer
+from .site_packages import hid
 
-import ctypes
-ctypes.CDLL("./hid_lib/hidapi.dll")
-
-from hid_lib import hid  # noqa
+from .MonitorDevice import MonitorDevice
+from .Console import Console
+from .Device import Device
 
 USAGE_PAGE = 0xFF31
 USAGE = 0x0074
-
-
-class Device(TypedDict):
-    vendor_id: int
-    product_id: int
-    serial_number: str
-    release_number: int
-    manufacturer_string: str
-    usage_page: int
-    usage: int
-    interface_number: int
-    bus_type: hid.BusType
-    index: int
-    thread: Thread
-    e: Exception
-    e_name: str
-    path: bytes
-
-
-class Console(ABC):
-    @staticmethod
-    def print_connect(device: Device) -> None:
-        pass
-
-    @staticmethod
-    def print_disconnect(device: Device) -> None:
-        pass
-
-    @staticmethod
-    def print_couldnt_connect(device: Device) -> None:
-        pass
-
-
-class MonitorDevice:
-    def __init__(self, hid_device: Device, observers: tuple[Observer] = ()):
-        self.hid_device = hid_device
-        self.device = hid.Device(path=hid_device["path"])
-        self.observers = observers
-        self.current_line = ""
-
-        Console.print_connect(hid_device)
-
-    def read(self, size, encoding="ascii", timeout=1) -> str:
-        """Read size bytes from the device."""
-        return self.device.read(size, timeout).decode(encoding)
-
-    def read_line(self) -> str:
-        """Read from the device's console until we get a \n."""
-        while "\n" not in self.current_line:
-            self.current_line += self.read(32).replace("\x00", "")
-
-        line, self.current_line = self.current_line.split("\n", 1)
-        return line
-
-    def read_message(self):
-        message = self.read_line()
-
-        for observer in self.observers:
-            observer.notify(message)
-
-    def run_forever(self):
-        while True:
-            try:
-                self.read_message()
-
-            except hid.HIDException:
-                break
 
 
 class DeviceListener:
