+++ companion/hid_listener/__init__.py
@@ -1,0 +1,2 @@
+
+from .DeviceListener import DeviceListener
