+++ companion/hid_listener/site_packages/__init__.py
@@ -1,0 +1,3 @@
+
+import ctypes
+ctypes.CDLL("./hid_listener/hidapi.dll")
