+++ companion/hid_listener/site_packages/hid/__init__.py
